+++ lib/src/indexeddb_store.dart
@@ -12,144 +12,133 @@
 //See the License for the specific language governing permissions and
 //limitations under the License.
 
-part of lawndart;
+part of '../lawndart.dart';
 
-/**
- * Wraps the IndexedDB API and exposes it as a [Store].
- * IndexedDB is generally the preferred API if it is available.
- */
+/// Wraps the IndexedDB API and exposes it as a [Store].
+/// IndexedDB is generally the preferred API if it is available.
 class IndexedDbStore extends Store {
-  static Map<String, idb.Database> _databases = new Map<String, idb.Database>();
+  IndexedDbStore._(this.dbName, this.storeName) : super._();
+  static final _databases = <String, IDBDatabase>{};
 
   final String dbName;
   final String storeName;
 
-  IndexedDbStore._(this.dbName, this.storeName) : super._();
-
   static Future<IndexedDbStore> open(String dbName, String storeName) async {
-    var store = new IndexedDbStore._(dbName, storeName);
+    final store = IndexedDbStore._(dbName, storeName);
     await store._open();
     return store;
   }
 
-  /// Returns true if IndexedDB is supported on this platform.
-  static bool get supported => idb.IdbFactory.supported;
-
-  Future _open() async {
-    if (!supported) {
-      throw new UnsupportedError('IndexedDB is not supported on this platform');
-    }
-
+  Future<bool> _open() async {
     final existingDb = _databases[dbName];
     if (existingDb != null) {
       existingDb.close();
     }
 
     final indexedDB = window.indexedDB;
-    if (indexedDB != null) {
-      var db = await indexedDB.open(dbName);
-      // print("Newly opened db $dbName has version ${db.version} and stores ${db.objectStoreNames}");
+    final request = indexedDB.open(dbName);
+    final completer = Completer<bool>();
+    request.onsuccess = () {
+      final db = request.result! as IDBDatabase;
+      // print("Newly opened db $dbName has version ${db.version}
+      // and stores ${db.objectStoreNames}");
       final objectStoreNames = db.objectStoreNames;
-      if (objectStoreNames != null && !objectStoreNames.contains(storeName)) {
+      if (!objectStoreNames.contains(storeName)) {
         db.close();
         // print('Attempting upgrading $storeName from ${db.version}');
-        db = await indexedDB.open(dbName, version: (db.version ?? 0) + 1,
-            onUpgradeNeeded: (e) {
-          // print('Upgrading db $dbName to ${db.version ?? 0 + 1}');
-          idb.Database d = e.target.result;
-          d.createObjectStore(storeName);
-        });
+        final upgradeRequest = indexedDB.open(dbName, db.version + 1);
+        upgradeRequest
+          ..onupgradeneeded = (() {
+            // print('Upgrading db $dbName to ${db.version ?? 0 + 1}');
+            (upgradeRequest.result! as IDBDatabase)
+                .createObjectStore(storeName);
+          }).toJS
+          ..onsuccess = () {
+            final upgradedDb = upgradeRequest.result! as IDBDatabase;
+            _databases[dbName] = upgradedDb;
+            completer.complete(true);
+          }.toJS;
+      } else {
+        _databases[dbName] = db;
+        completer.complete(true);
       }
-
-      _databases[dbName] = db;
-      return true;
-    }
-    return false;
+    }.toJS;
+    return completer.future;
   }
 
-  idb.Database get _db => _databases[dbName]!;
+  IDBDatabase get _db => _databases[dbName]!;
 
   @override
-  Future removeByKey(String key) {
-    return _runInTxn((store) => store.delete(key));
+  Future<void> removeByKey(String key) =>
+      _runInTxn((store) => store.delete(key.toJS));
+
+  @override
+  Future<String> save(String obj, String key) =>
+      _runInTxn<String>((store) => store.put(obj.toJS, key.toJS));
+
+  @override
+  Future<String?> getByKey(String key) =>
+      _runInTxn<String?>((store) => store.get(key.toJS), 'readonly');
+
+  @override
+  Future<void> nuke() => _runInTxn((store) => store.clear());
+
+  Future<T> _runInTxn<T>(
+    IDBRequest Function(IDBObjectStore store) requestCommand, [
+    String txnMode = 'readwrite',
+  ]) async {
+    final trans = _db.transaction(storeName.toJS, txnMode);
+    final store = trans.objectStore(storeName);
+    final request = requestCommand(store);
+    final completer = Completer<T>();
+    trans.oncomplete = () {
+      completer.complete(request.result.dartify() as T);
+    }.toJS;
+    return completer.future;
+  }
+
+  Stream<String> _doGetAll(
+    String Function(IDBCursorWithValue cursor) onCursor,
+  ) {
+    final trans = _db.transaction(storeName.toJS, 'readonly');
+    final store = trans.objectStore(storeName);
+    final request = store.openCursor();
+    // ignore: close_sinks
+    final controller = StreamController<String>();
+    request.onsuccess = () {
+      if (request.result case final IDBCursorWithValue cursor) {
+        final key = onCursor(cursor);
+        controller.add(key);
+        cursor.continue_();
+      } else {
+        unawaited(controller.sink.close());
+      }
+    }.toJS;
+    return controller.stream;
   }
 
   @override
-  Future<String> save(String obj, String key) {
-    return _runInTxn<String>(
-        (store) async => (await store.put(obj, key)) as String);
-  }
-
-  @override
-  Future<String?> getByKey(String key) {
-    return _runInTxn<String?>(
-        (store) async => (await store.getObject(key)) as String?, 'readonly');
-  }
-
-  @override
-  Future nuke() {
-    return _runInTxn((store) => store.clear());
-  }
-
-  Future<T> _runInTxn<T>(Future<T> requestCommand(idb.ObjectStore store),
-      [String txnMode = 'readwrite']) async {
-    var trans = _db.transaction(storeName, txnMode);
-    var store = trans.objectStore(storeName);
-    var result = await requestCommand(store);
-    await trans.completed;
-    return result;
-  }
-
-  Stream<String> _doGetAll(String onCursor(idb.CursorWithValue cursor)) async* {
-    var trans = _db.transaction(storeName, 'readonly');
-    var store = trans.objectStore(storeName);
-    await for (var cursor in store.openCursor(autoAdvance: true)) {
-      yield onCursor(cursor);
-    }
-  }
-
-  @override
-  Stream<String> all() {
-    return _doGetAll((idb.CursorWithValue cursor) => cursor.value);
-  }
-
-  @override
-  Future batch(Map<String, String> objs) {
-    return _runInTxn((store) {
-      objs.forEach((k, v) {
-        store.put(v, k);
-      });
-      return Future.value();
-    });
-  }
+  Stream<String> all() =>
+      _doGetAll((cursor) => (cursor.value! as JSString).toDart);
 
   @override
   Stream<String> getByKeys(Iterable<String> keys) async* {
-    for (var key in keys) {
-      var v = await getByKey(key);
-      if (v != null) yield v;
-    }
-  }
-
-  @override
-  Future<bool> removeByKeys(Iterable<String> keys) {
-    return _runInTxn((store) {
-      for (var key in keys) {
-        store.delete(key);
+    for (final key in keys) {
+      final v = await getByKey(key);
+      if (v != null) {
+        yield v;
       }
-      return Future.value();
-    });
+    }
   }
 
   @override
   Future<bool> exists(String key) async {
-    var value = await getByKey(key);
+    final value = await getByKey(key);
     return value != null;
   }
 
+  // works as long as nothing else writes non-String keys
   @override
-  Stream<String> keys() {
-    // works as long as nothing else writes non-String keys
-    return _doGetAll((idb.CursorWithValue cursor) => cursor.key! as String);
-  }
+  Stream<String> keys() =>
+      _doGetAll((cursor) => (cursor.key! as JSString).toDart);
 }